pyproject.toml
[tool.poetry]
name = "datalake-gcp-demo"
version = "0.1.0"
description = ""
authors = ["LeeoSilva <leeosilva0909@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
diagrams = "^0.23.4"
apache-airflow = {version="^2.9.2", extras=["google"]}
paramiko = "^3.4.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
duckdb = "^1.0.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
